# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/modules": {
      "post": {
        "description": "Creates a new module record.",
        "operationId": "ModulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created module.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Module"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a new module",
        "tags": [
          "modules"
        ]
      },
      "get": {
        "description": "Returns a list of all modules.",
        "operationId": "ModulesController_findAll",
        "parameters": [
          {
            "name": "minQuestions",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxQuestions",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "exactQuestions",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "universityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nameSearch",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasQuestions",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasPrerequisites",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasPostrequisites",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasSubModules",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasParentModules",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "fewQuestions",
            "required": false,
            "in": "query",
            "description": "Filter modules with fewer than 20 questions",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of modules.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Module"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve all modules",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/bulk": {
      "post": {
        "description": "Creates multiple module records in a single operation.",
        "operationId": "ModulesController_createMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManyModulesInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The number of modules created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number",
                      "description": "Number of modules created",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create multiple modules",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/questions-data": {
      "get": {
        "description": "Returns all modules with id, en_name, and questions_amount, sorted by question count (least questions first).",
        "operationId": "ModulesController_getModulesQuestionsData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A list of modules with their question counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "en_name": {
                        "type": "string"
                      },
                      "questions_amount": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Get modules questions data",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/{id}": {
      "get": {
        "description": "Returns a single module by its ID.",
        "operationId": "ModulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the module",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The module with the specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Module"
                }
              }
            }
          },
          "404": {
            "description": "Module not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve a module by ID",
        "tags": [
          "modules"
        ]
      },
      "put": {
        "description": "Updates an existing module record.",
        "operationId": "ModulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the module",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated module.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Module"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Module not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Update a module by ID",
        "tags": [
          "modules"
        ]
      },
      "delete": {
        "description": "Deletes a module record by its ID.",
        "operationId": "ModulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the module",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Module successfully deleted."
          },
          "404": {
            "description": "Module not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a module by ID",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/relationship": {
      "post": {
        "description": "Creates a new relationship between modules.",
        "operationId": "ModulesController_createRelationship",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleRelationshipInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created module relationship.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRelationshipResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a module relationship",
        "tags": [
          "modules"
        ]
      },
      "delete": {
        "description": "Deletes an existing relationship between modules.",
        "operationId": "ModulesController_deleteRelationship",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteModuleRelationshipInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Module relationship successfully deleted."
          },
          "404": {
            "description": "Module relationship not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a module relationship",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/{id}/summary": {
      "get": {
        "description": "Returns a human-readable plain text summary for the specified module.",
        "operationId": "ModulesController_getSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the module",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plain text summary of the module",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Module not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Get module summary",
        "tags": [
          "modules"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumina API",
    "description": "API documentation for Lumina server",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Translation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "en_text": {
            "type": "string"
          },
          "he_text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "en_text",
          "he_text"
        ]
      },
...
      "RelationshipMetadataKey": {
        "type": "string",
        "enum": [
          "REASON",
          "TYPE",
          "DESCRIPTION"
        ]
      },
      "RelationshipMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RelationshipMetadataKey"
              }
            ]
          },
          "value": {
            "type": "string"
          },
          "blockRelationshipId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "value",
          "blockRelationshipId"
        ]
      },
      "BlockRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteId": {
            "type": "string"
          },
          "postrequisiteId": {
            "type": "string"
          },
          "prerequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "postrequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "metadata": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationshipMetadata"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteId",
          "postrequisiteId",
          "prerequisite",
          "postrequisite",
          "metadata"
        ]
      },
      "Block": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteFor": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "postrequisiteOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteFor",
          "postrequisiteOf",
          "modules"
        ]
      },
      "QuestionValidationStatus": {
        "type": "string",
        "enum": [
          "ai_generated",
          "in_manual_review",
          "approved",
          "rejected"
        ]
      },
      "SelectAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCorrect",
          "translationId",
          "text",
          "answerId"
        ]
      },
      "UnitAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "unit",
          "answerId"
        ]
      },
      "NumberAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "answerId"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "SelectAnswer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectAnswer"
            }
          },
          "UnitAnswer": {
            "$ref": "#/components/schemas/UnitAnswer"
          },
          "NumberAnswer": {
            "$ref": "#/components/schemas/NumberAnswer"
          }
        },
        "required": [
          "id",
          "questionId"
        ]
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "selection",
          "value",
          "void"
        ]
      },
      "QuestionPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "partQuestionId": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "questionId",
          "partQuestionId",
          "order"
        ]
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "validationStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionValidationStatus"
              }
            ]
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "Answer": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Answer"
            }
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              }
            ]
          },
          "Modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Parts": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          },
          "PartOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          }
        },
        "required": [
          "id",
          "validationStatus",
          "translationId",
          "text",
          "Answer",
          "type",
          "Modules",
          "Parts",
          "PartOf"
        ]
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "translationId": {
            "type": "string"
          },
          "subModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "parentModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "questions": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        },
        "required": [
          "id",
          "name",
          "translationId",
          "subModules",
          "parentModules",
          "Block",
          "questions"
        ]
      },
...
      "CreateModuleInput": {
        "type": "object",
        "properties": {
          "en_text": {
            "type": "string",
            "description": "English name of the module"
          },
          "he_text": {
            "type": "string",
            "description": "Hebrew name of the module"
          },
          "courseId": {
            "type": "string",
            "description": "ID of the course this module belongs to"
          }
        },
        "required": [
          "en_text",
          "he_text",
          "courseId"
        ]
      },
      "CreateManyModulesInput": {
        "type": "object",
        "properties": {
          "modules": {
            "description": "Array of module data to create",
            "example": [
              {
                "en_text": "Linear Algebra",
                "he_text": "אלגברה לינארית",
                "courseId": "uuid1"
              },
              {
                "en_text": "Calculus",
                "he_text": "חשבון אינפיניטסימלי",
                "courseId": "uuid2"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateModuleInput"
            }
          }
        },
        "required": [
          "modules"
        ]
      },
      "CreateModuleRelationshipInput": {
        "type": "object",
        "properties": {
          "prerequisiteModuleId": {
            "type": "string",
            "description": "The ID of the prerequisite module"
          },
          "postrequisiteModuleId": {
            "type": "string",
            "description": "The ID of the postrequisite module"
          },
          "metadata": {
            "type": "object",
            "description": "Optional metadata for the relationship (JSON string)"
          }
        },
        "required": [
          "prerequisiteModuleId",
          "postrequisiteModuleId"
        ]
      },
      "ModuleRelationshipResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisite": {
            "$ref": "#/components/schemas/Block"
          },
          "postrequisite": {
            "$ref": "#/components/schemas/Block"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "prerequisite",
          "postrequisite",
          "metadata"
        ]
      },
      "DeleteModuleRelationshipInput": {
        "type": "object",
        "properties": {
          "prerequisiteModuleId": {
            "type": "string",
            "description": "The ID of the prerequisite module"
          },
          "postrequisiteModuleId": {
            "type": "string",
            "description": "The ID of the postrequisite module"
          }
        },
        "required": [
          "prerequisiteModuleId",
          "postrequisiteModuleId"
        ]
      },
...
    }
  }
}